Put ChangeLogs in a list instead of map
Originally, the changelogs are put in a map, with commit message as the
key and author as the value, that later will be exposed as JSON. If
there is a commit message that contains double quotes, the resulting
JSON will have unescaped double quotes in the JSON object keys and thus
makes the JSON invalid. The Pipeline Aggregator web view will fail to
parse it and stop showing the builds.

This commit changes the changelogs to use list of a structure instead of
a simple map. The resulting JSON will have properly escaped commit
messages.